Extract case sensitivity probing partial
