+++ src/CodeIndex/Indexer/Scanning/FileIndexer.cs
@@ -352,112 +352,6 @@ public partial class FileIndexer
         (_submodulePaths, _submoduleAncestorPaths, _submoduleLoadWarnings) = LoadGitSubmodulePaths(_ignoreRuleRoot, _projectRoot, pathComparer);
     }
 
-    private static bool ProbeFileSystemIgnoreCase(string projectRoot)
-    {
-        var normalizedRoot = projectRoot;
-        try
-        {
-            normalizedRoot = Path.GetFullPath(projectRoot);
-            if (FileSystemIgnoreCaseProbeForTesting is { } probeOverride)
-                return probeOverride(normalizedRoot);
-
-            if (TryProbeExistingDirectoryPath(normalizedRoot, out var ignoreCase))
-                return ignoreCase;
-
-            using var probe = CaseSensitivityProbeDirectory.CreateProbePathScope(normalizedRoot, "case-probe-");
-            var probePath = probe.Path;
-            FileWriteProbe.WriteEmptyFile(probePath);
-            try
-            {
-                if (TryCreateCaseVariant(probePath, out var probeVariant))
-                    return File.Exists(LongPath.EnsureWindowsPrefix(probeVariant));
-            }
-            finally
-            {
-                FileWriteProbe.DeleteFileIfExists(probePath);
-            }
-
-            throw new CaseSensitivityProbeException(
-                "Failed to create a case-variant path for filesystem case-sensitivity probing.",
-                normalizedRoot,
-                probePath: probePath);
-        }
-        catch (CaseSensitivityProbeException)
-        {
-            throw;
-        }
-        catch (Exception ex) when (IsCaseSensitivityProbeFailure(ex))
-        {
-            throw new CaseSensitivityProbeException(
-                "Failed to determine filesystem case sensitivity.",
-                normalizedRoot,
-                ex);
-        }
-    }
-
-    private static bool TryProbeExistingDirectoryPath(string path, out bool ignoreCase)
-    {
-        ignoreCase = false;
-        if (!TryCreateCaseVariant(path, out var variant))
-            return false;
-
-        ignoreCase = Directory.Exists(LongPath.EnsureWindowsPrefix(variant));
-        return true;
-    }
-
-    private static bool IsCaseSensitivityProbeFailure(Exception ex)
-        => ex is ArgumentException
-            or IOException
-            or UnauthorizedAccessException
-            or NotSupportedException
-            or System.Security.SecurityException;
-
-    private static bool TryCreateCaseVariant(string path, out string variant)
-    {
-        var chars = path.ToCharArray();
-        for (var i = chars.Length - 1; i >= 0; i--)
-        {
-            var ch = chars[i];
-            if (!char.IsLetter(ch))
-                continue;
-
-            chars[i] = char.IsUpper(ch)
-                ? char.ToLowerInvariant(ch)
-                : char.ToUpperInvariant(ch);
-            variant = new string(chars);
-            return true;
-        }
-
-        variant = path;
-        return false;
-    }
-
-    private static bool? ProbeExistingDirectoryIgnoreCase(string directory)
-    {
-        try
-        {
-            var normalizedDirectory = Path.GetFullPath(directory);
-            return TryCreateCaseVariant(normalizedDirectory, out var variant)
-                ? Directory.Exists(LongPath.EnsureWindowsPrefix(variant))
-                : null;
-        }
-        catch
-        {
-            return null;
-        }
-    }
-
-    private bool DirectoryUsesIgnoreCase(string directory)
-    {
-        var fullPath = Path.GetFullPath(directory);
-        if (_directoryIgnoreCaseCache.TryGetValue(fullPath, out var ignoreCase))
-            return ignoreCase;
-
-        ignoreCase = _directoryIgnoreCaseProbe(fullPath) ?? _ignoreCase;
-        _directoryIgnoreCaseCache[fullPath] = ignoreCase;
-        return ignoreCase;
-    }
-
     /// <summary>
     /// Return all file patterns (extensions and filenames) mapped to their language names.
     /// 全ファイルパターン（拡張子とファイル名）と対応する言語名のマッピングを返す。
